refactor: reorganize features in folders
