+++ src/server/features/go-definition/index.ts
@@ -1,0 +1,1 @@
+export * from "./go-definition";
