+++ src/server/features/workspace-symbols/index.ts
@@ -1,0 +1,1 @@
+export * from "./workspace-symbol";
