+++ src/test/features/completion.spec.ts
@@ -4,8 +4,8 @@ import type { CompletionList } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { doCompletion } from "../../server/features/completion";
 import { parseStringLiteralChoices } from "../../server/features/completion/completion-utils";
+import { sassBuiltInModules } from "../../server/features/sass-built-in-modules";
 import { ScssDocument } from "../../server/parser";
-import { sassBuiltInModules } from "../../server/sass-built-in-modules";
 import type { ISettings } from "../../server/settings";
 import StorageService from "../../server/storage";
 import * as helpers from "../helpers";
