+++ src/test/features/hover.spec.ts
@@ -2,7 +2,7 @@ import assert from "assert";
 import { MarkupKind, SymbolKind } from "vscode-languageserver";
 import type { Hover } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
-import { doHover } from "../../server/features/hover";
+import { doHover } from "../../server/features/hover/hover";
 import { ScssDocument } from "../../server/parser";
 import StorageService from "../../server/storage";
 import * as helpers from "../helpers";
