+++ src/test/features/signature-help.spec.ts
@@ -2,7 +2,7 @@ import assert from "assert";
 import { SymbolKind } from "vscode-languageserver";
 import type { SignatureHelp } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
-import { doSignatureHelp } from "../../server/features/signature-help";
+import { doSignatureHelp } from "../../server/features/signature-help/signature-help";
 import { ScssDocument } from "../../server/parser";
 import StorageService from "../../server/storage";
 import * as helpers from "../helpers";
